requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.2.0
astunparse==1.6.3
attrs==23.2.0
bcrypt==4.1.2
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
ecdsa==0.18.0
exceptiongroup==1.2.1
fastapi==0.109.2
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.5.4
google-api-core==2.19.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
h5py==3.11.0
idna==3.6
jax==0.4.28
jaxlib==0.4.28
keras==2.15.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.2.0
namex==0.0.8
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
passlib==1.7.4
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.18.0
PyMySQL==1.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.9
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scipy==1.13.0
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
SQLAlchemy==2.0.27
starlette==0.36.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing_extensions==4.12.0
urllib3==2.2.0
uvicorn==0.27.1
Werkzeug==3.0.3
wrapt==1.14.1
